processing-service/requirements.txt
boto3==1.35.*
pillow==10.*
mtcnn==0.1.1
tensorflow-cpu==2.15.*
opencv-python==4.10.*
numpy
